feat: created a settingsRow component
